added / updated routes, responses, controllers and model for loans

+++ mb40mktg/database/migrations/2018_09_12_133356_modify_loan_table.php
@@ -32,7 +32,7 @@ class ModifyLoanTable extends Migration
         if (Schema::hasTable($this->tableName)) {
             Schema::table($this->tableName, function (Blueprint $table) {
                 $table->integer("item_id");
-                $table->integer("interest_rt");
+                $table->float("interest_rt");
             });
         }
     }
